PiOLED option changes (#1030)

+++ mycodo/functions/display_ssd1306_oled_128x64_spi.py
@@ -264,7 +264,7 @@ FUNCTION_INFORMATION = {
         },
         {
             'id': 'font_size',
-            'type': 'float',
+            'type': 'integer',
             'default_value': 10,
             'required': True,
             'constraints_pass': constraints_pass_positive_value,
